+++ admin/view/javascript/nuvei_orders.js
@@ -8,22 +8,32 @@ nuveiVars.ajaxUrl = 'index.php?route=' + nuveiControllerPath + '&' + nuveiTokenN
 
 function scOrderActions(confirmQusetion, action, orderId) {
 	console.log(action);
+    
+    var reqData = {
+        orderId: orderId
+        ,action: action
+    };
 
 	if('refund' == action) {
-		var refAm	= $('#refund_amount').val().replace(',', '.');
+		var refAm   = $('#refund_amount').val().replace(',', '.');
 		var reg		= new RegExp(/^\d+(\.\d{1,2})?$/); // match integers and decimals
 
-		if(action == 'refund' || action == 'refundManual') { 
-			if(!reg.test(refAm) || isNaN(refAm) || refAm <= 0) {
-				alert(nuveiVars.nuveiRefundAmountError);
-				return;
-			}
-		}
+        if(!reg.test(refAm) || isNaN(refAm) || refAm <= 0) {
+            alert(nuveiVars.nuveiRefundAmountError);
+            return;
+        }
+        
+        reqData.amount = refAm;
 	}
+    
+    if ('cancelSubscr' != action) {
+        confirmQusetion += ' #' + orderId;
+    }
+    
+    confirmQusetion += '?';
 
-	if(confirm(confirmQusetion + ' #' + orderId + '?')) {
-		var spinnerId = (action == 'void' ? 'void' : 'refund') + '_spinner';
-		$('#' + spinnerId).removeClass('hide');
+	if(confirm(confirmQusetion)) {
+		$('#nuvei_spinner').removeClass('hide');
 		
 		// disable sc custom buttons
 		$('.sc_order_btns').each(function(){
@@ -36,14 +46,10 @@ function scOrderActions(confirmQusetion, action, orderId) {
 			url: nuveiVars.ajaxUrl,
 			type: 'post',
 			dataType: 'json',
-			data: {
-				orderId: orderId
-				,action: action
-				,amount: $('#refund_amount').val()
-			}
+			data: reqData
 		})
 		.done(function(resp) {
-			console.log('done', resp)
+			console.log('done', resp);
 	
 			if(resp.hasOwnProperty('status')) {
 				if(resp.status == 1) {
@@ -60,7 +66,7 @@ function scOrderActions(confirmQusetion, action, orderId) {
 					}
 					
 
-					$('#' + spinnerId).addClass('hide');
+					$('#nuvei_spinner').addClass('hide');
 					
 					// enable sc custom buttons
 					$('.sc_order_btns').each(function(){
@@ -115,50 +121,46 @@ function deleteManualRefund(id, amount, orderId) {
 
 function loadNuveiExtras() {
 	// 1.set the changes in Options table
-	var scPlaceOne          = $('#content .container-fluid .row .col-md-4:nth-child(3)').find('table tbody');
-    var scRefundBtnsHtml    = '';
-    var scVoidSettleButton  = '';
+	var scPlaceOne      = $('#content .container-fluid .row .col-md-4:nth-child(3)').find('table tbody');
+    var nuveiButtons    = '';
 
 	if(scPlaceOne.length > 0) {
-		// 1.1.place Refund and Void buttons
-		if(1 == nuveiVars.nuveiAllowRefundBtn) {
-			scRefundBtnsHtml = 
-                '<div class="input-group" style="margin-bottom: 2px; margin-top: 2px;">'
-                    + '<input type="text" class="form-control" id="refund_amount" value="" style="max-height: 22px;">'
+        nuveiButtons +=
+            '<div class="input-group pull-right" style="">';
+    
+        if('1' == nuveiVars.nuveiAllowSettleBtn) {
+			nuveiButtons += '<button class="btn btn-success btn-xs sc_order_btns" style="margin-bottom: 2px; margin-left: 2px;" onclick="scOrderActions(\''+ nuveiVars.nuveiOrderConfirmSettle +'\', \'settle\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.nuveiBtnSettle +'</button>';
+		}
+    
+        if(1 == nuveiVars.nuveiAllowVoidBtn) {
+            nuveiButtons += '<button class="btn btn-danger btn-xs sc_order_btns" style="margin-bottom: 2px; margin-left: 2px;" onclick="scOrderActions(\''+ nuveiVars.nuveiOrderConfirmCancel +'\', \'void\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.nuveiBtnVoid +'</button>';
+        }
+        
+        if (1 == nuveiVars.nuveiAllowCancelSubsBtn) {
+            nuveiButtons += '<button class="btn btn-danger btn-xs sc_order_btns" style="margin-bottom: 2px; margin-left: 2px;" onclick="scOrderActions(\''+ nuveiVars.orderConfirmCancelSubscr +'\', \'cancelSubscr\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.btnCancelSubscr +'</button>';
+        }
+        
+        nuveiButtons +=
+            '</div>';
+
+        if(1 == nuveiVars.nuveiAllowRefundBtn) {
+            nuveiButtons +=
+                '<div class="input-group pull-right" style="display: inline-block; margin-bottom: 2px;">'
+                    + '<input type="text" class="form-control" id="refund_amount" value="" style="max-height: 22px; max-width: 70px;">'
                     + '<span class="input-group-btn">'
-                        + '<button class="btn btn-danger sc_order_btns btn-xs" type="button" onclick="scOrderActions(\''+ nuveiVars.nuveiOrderConfirmRefund +'\', \'refund\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.nuveiBtnRefund +'</button>'
-                    + '</span>';
-            
-            // add Void button
-            if(1 == nuveiVars.nuveiAllowVoidBtn) {
-                scRefundBtnsHtml +=
-                    '<span class="input-group-btn" style="left: 2px;">'
-                        + '<button class="btn btn-danger btn-xs sc_order_btns" style="margin-left: 2px;" onclick="scOrderActions(\''+ nuveiVars.nuveiOrderConfirmCancel +'\', \'void\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.nuveiBtnVoid +'</button>'
-                    + '</span>';
-            }
-            
-            scRefundBtnsHtml +=
-                '</div>';
-		}
-		// 1.2.set Void button only
-		else if(1 == nuveiVars.nuveiAllowVoidBtn) {
-			scVoidSettleButton += '<button class="btn btn-danger btn-xs sc_order_btns" style="margin-left: 2px;" onclick="scOrderActions(\''+ nuveiVars.nuveiOrderConfirmCancel +'\', \'void\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.nuveiBtnVoid +'</button>';
-		}
-
-		// 1.3.set Settle btn
-		if('1' == nuveiVars.nuveiAllowSettleBtn) {
-			scVoidSettleButton += '<button class="btn btn-success btn-xs sc_order_btns" style="margin-left: 2px;" onclick="scOrderActions(\''+ nuveiVars.nuveiOrderConfirmSettle +'\', \'settle\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.nuveiBtnSettle +'</button>';
-		}
-
-		// 1.4.place buttons
-		scPlaceOne.append(
+                        + '<button class="btn btn-danger sc_order_btns btn-xs pull-right" type="button" onclick="scOrderActions(\''+ nuveiVars.nuveiOrderConfirmRefund +'\', \'refund\', '+ nuveiVars.nuveiOrderId +')">'+ nuveiVars.nuveiBtnRefund +'</button>'
+                    + '</span>'
+                    
+                + '</div>';
+        }
+        
+        scPlaceOne.append(
 			'<tr>'
 				+ '<td>'
-					+ '<span>'+ nuveiVars.nuveiMoreActions +'&nbsp;&nbsp;<i id="void_spinner" class="fa fa-circle-o-notch fa-spin hide"></i></span>'
+					+ '<span>'+ nuveiVars.nuveiMoreActions +'&nbsp;&nbsp;<i id="nuvei_spinner" class="fa fa-circle-o-notch fa-spin hide"></i></span>'
 				+ '</td>'
 				+ '<td colspan="2" class="text-right col-xs-6">'
-                    + scRefundBtnsHtml
-                    + ('' != scVoidSettleButton ? '<div class="btn-group" role="group">' + scVoidSettleButton + '</div>' : '')
+                    + '<div class="d-flex">' + nuveiButtons + '</div>'
 				+ '</td>'
 			+ '</tr>'
 		);