+++ system/library/nuvei/NUVEI_CLASS.php
@@ -8,31 +8,27 @@
  * @author Nuvei
  */
 
-define('NUVEI_PLUGIN_V',            '1.1');
-define('NUVEI_PLUGIN_CODE',         'nuvei');
-define('NUVEI_PLUGIN_TITLE',        'Nuvei');
+const NUVEI_PLUGIN_V            = '1.2';
+const NUVEI_PLUGIN_CODE         = 'nuvei';
+const NUVEI_PLUGIN_TITLE        = 'Nuvei';
 
-define('NUVEI_LIVE_URL_BASE',       'https://secure.safecharge.com/ppp/api/v1/');
-define('NUVEI_TEST_URL_BASE',       'https://ppp-test.safecharge.com/ppp/api/v1/');
+const NUVEI_LIVE_URL_BASE       = 'https://secure.safecharge.com/ppp/api/v1/';
+const NUVEI_TEST_URL_BASE       = 'https://ppp-test.safecharge.com/ppp/api/v1/';
+const NUVEI_SDK_URL_INT         = 'https://srv-bsf-devpppjs.gw-4u.com/checkoutNext/checkout.js';
+const NUVEI_SDK_URL_PROD        = 'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js';
+const NUVEI_SDK_AUTOCLOSE_URL   = 'https://cdn.safecharge.com/safecharge_resources/v1/websdk/autoclose.html';
 
-define('NUVEI_SDK_URL_INT',         'https://srv-bsf-devpppjs.gw-4u.com/checkoutNext/checkout.js');
-define('NUVEI_SDK_URL_PROD',        'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js');
-
-define('NUVEI_SDK_AUTOCLOSE_URL',   'https://cdn.safecharge.com/safecharge_resources/v1/websdk/autoclose.html');
-
-define('NUVEI_AUTH_CODE',           '_authCode');
-define('NUVEI_TRANS_ID',            '_transactionId');
-define('NUVEI_TRANS_TYPE',          '_transactionType');
+const NUVEI_AUTH_CODE           = '_authCode';
+const NUVEI_TRANS_ID            = '_transactionId';
+const NUVEI_TRANS_TYPE          = '_transactionType';
 
 // if change the both consts above, change the ajaxURL in admin nuvei_order.js
-define('NUVEI_TOKEN_NAME',          'user_token');
-define('NUVEI_CONTROLLER_PATH',     'extension/payment/nuvei');
-
-define('NUVEI_SETTINGS_PREFIX',     'payment_nuvei_');
-define('NUVEI_SOURCE_APP',          'OPENCART_3_0_PLUGIN');
-
-//define('NUVEI_ADMIN_TXT_EXT_KEY',   'text_extension');
-define('NUVEI_ADMIN_EXT_URL',       'marketplace/extension');
+const NUVEI_TOKEN_NAME          = 'user_token';
+const NUVEI_CONTROLLER_PATH     = 'extension/payment/nuvei';
+const NUVEI_SETTINGS_PREFIX     = 'payment_nuvei_';
+const NUVEI_SOURCE_APP          = 'OPENCART_3_0_PLUGIN';
+//const NUVEI_ADMIN_TXT_EXT_KEY = 'text_extension';
+const NUVEI_ADMIN_EXT_URL       = 'marketplace/extension';
 
 class NUVEI_CLASS
 {