+++ system/library/nuvei/ocmod_includes/admin_header.php
@@ -1,0 +1,5 @@
+<?php
+
+if($this->user->isLogged()) {
+    $this->document->addScript('view/javascript/nuvei_version_checker.js');
+}